my_coding_agent/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.0.0
gitpython>=3.1.0
